refactor(node): decompose main_tests sidecar into modules (#2750)
